migrated interfaces out of a file to satisfy max file length
