+++ AzureCommunicationUI/sdk/AzureCommunicationUICalling/Sources/Presentation/Factories/CompositeViewModelFactory.swift
@@ -6,114 +6,6 @@
 import FluentUI
 import Foundation
 
-// swiftlint:disable file_length
-protocol CompositeViewModelFactoryProtocol {
-    // MARK: CompositeViewModels
-    func getSetupViewModel() -> SetupViewModel
-
-    func getCallingViewModel() -> CallingViewModel
-
-    // MARK: ComponentViewModels
-    func makeIconButtonViewModel(iconName: CompositeIcon,
-                                 buttonType: IconButtonViewModel.ButtonType,
-                                 isDisabled: Bool,
-                                 action: @escaping (() -> Void)) -> IconButtonViewModel
-
-    func makeIconWithLabelButtonViewModel<ButtonStateType>(
-        selectedButtonState: ButtonStateType,
-        localizationProvider: LocalizationProviderProtocol,
-        buttonTypeColor: IconWithLabelButtonViewModel<ButtonStateType>.ButtonTypeColor,
-        isDisabled: Bool,
-        action: @escaping (() -> Void)) -> IconWithLabelButtonViewModel<ButtonStateType>
-
-    func makeLocalVideoViewModel(dispatchAction: @escaping ActionDispatch) -> LocalVideoViewModel
-
-    func makePrimaryButtonViewModel(buttonStyle: FluentUI.ButtonStyle,
-                                    buttonLabel: String,
-                                    iconName: CompositeIcon?,
-                                    isDisabled: Bool,
-                                    paddings: CompositeButton.Paddings?,
-                                    action: @escaping (() -> Void)) -> PrimaryButtonViewModel
-
-    func makeAudioDevicesListViewModel(dispatchAction: @escaping ActionDispatch,
-                                       localUserState: LocalUserState) -> AudioDevicesListViewModel
-
-    func makeErrorInfoViewModel(title: String,
-                                subtitle: String) -> ErrorInfoViewModel
-
-    func makeCallDiagnosticsViewModel(dispatchAction: @escaping ActionDispatch) -> CallDiagnosticsViewModel
-
-    // MARK: CallingViewModels
-    func makeLobbyOverlayViewModel() -> LobbyOverlayViewModel
-
-    func makeLoadingOverlayViewModel() -> LoadingOverlayViewModel
-
-    func makeOnHoldOverlayViewModel(resumeAction: @escaping (() -> Void)) -> OnHoldOverlayViewModel
-
-    func makeControlBarViewModel(dispatchAction: @escaping ActionDispatch,
-                                 endCallConfirm: @escaping (() -> Void),
-                                 localUserState: LocalUserState) -> ControlBarViewModel
-    func makeInfoHeaderViewModel(dispatchAction: @escaping ActionDispatch,
-                                 localUserState: LocalUserState) -> InfoHeaderViewModel
-    func makeLobbyWaitingHeaderViewModel(localUserState: LocalUserState,
-                                         dispatchAction: @escaping ActionDispatch) -> LobbyWaitingHeaderViewModel
-    func makeLobbyActionErrorViewModel(localUserState: LocalUserState,
-                                       dispatchAction: @escaping ActionDispatch) -> LobbyErrorHeaderViewModel
-    func makeParticipantCellViewModel(participantModel: ParticipantInfoModel,
-                                      lifeCycleState: LifeCycleState) -> ParticipantGridCellViewModel
-    func makeParticipantGridsViewModel(isIpadInterface: Bool) -> ParticipantGridViewModel
-    func makeParticipantsListViewModel(localUserState: LocalUserState,
-                                       dispatchAction: @escaping ActionDispatch) -> ParticipantsListViewModel
-    func makeBannerViewModel() -> BannerViewModel
-
-    func makeBannerTextViewModel() -> BannerTextViewModel
-
-    func makeLocalParticipantsListCellViewModel(localUserState: LocalUserState) -> ParticipantsListCellViewModel
-
-    func makeParticipantsListCellViewModel(participantInfoModel: ParticipantInfoModel) -> ParticipantsListCellViewModel
-
-    func makeMoreCallOptionsListViewModel(showSharingViewAction: @escaping () -> Void,
-                                          showSupportFormAction: @escaping () -> Void) -> MoreCallOptionsListViewModel
-
-    func makeDebugInfoSharingActivityViewModel() -> DebugInfoSharingActivityViewModel
-
-    func makeDrawerListItemViewModel(icon: CompositeIcon,
-                                     title: String,
-                                     accessibilityIdentifier: String,
-                                     action: @escaping (() -> Void)) -> DrawerListItemViewModel
-
-    func makeSelectableDrawerListItemViewModel(
-        icon: CompositeIcon,
-        title: String,
-        isSelected: Bool,
-        onSelectedAction: @escaping (() -> Void)) -> SelectableDrawerListItemViewModel
-
-    func makeSupportFormViewModel() -> SupportFormViewModel
-
-    // MARK: SetupViewModels
-    func makePreviewAreaViewModel(dispatchAction: @escaping ActionDispatch) -> PreviewAreaViewModel
-
-    func makeSetupControlBarViewModel(dispatchAction: @escaping ActionDispatch,
-                                      localUserState: LocalUserState) -> SetupControlBarViewModel
-
-    func makeJoiningCallActivityViewModel() -> JoiningCallActivityViewModel
-}
-
-extension CompositeViewModelFactoryProtocol {
-    func makePrimaryButtonViewModel(buttonStyle: FluentUI.ButtonStyle,
-                                    buttonLabel: String,
-                                    iconName: CompositeIcon? = .none,
-                                    isDisabled: Bool = false,
-                                    action: @escaping (() -> Void)) -> PrimaryButtonViewModel {
-        return makePrimaryButtonViewModel(buttonStyle: buttonStyle,
-                                   buttonLabel: buttonLabel,
-                                   iconName: iconName,
-                                   isDisabled: isDisabled,
-                                   paddings: nil,
-                                   action: action)
-    }
-}
-
 class CompositeViewModelFactory: CompositeViewModelFactoryProtocol {
     private let logger: Logger
     private let store: Store<AppState, Action>
